-feature :sparkles: docked player money completed

+++ modules/payment/paymentModel.go
@@ -17,11 +17,11 @@ type (
 	}
 
 	PaymentTransferRes struct {
-		InventoryId   string `json:"inventory_id"`
-		TransactionId string `json:"transaction_id"`
-		PlayerId      string `json:"player_id"`
-		ItemId        string `json:"item_id"`
-		Amount        int64  `json:"amount"`
-		Error         string `json:"error"`
+		InventoryId   string  `json:"inventory_id"`
+		TransactionId string  `json:"transaction_id"`
+		PlayerId      string  `json:"player_id"`
+		ItemId        string  `json:"item_id"`
+		Amount        float64 `json:"amount"`
+		Error         string  `json:"error"`
 	}
 )
